requirements.txt

numpy
pandas
Pillow
streamlit
joblib
scikit-learn